% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map}
\alias{map}
\title{Map of any shape file overlayed with the world heritage site polygon and a map of South Africa}
\usage{
map(data, col, name)
}
\arguments{
\item{data}{The shape file to be mapped}

\item{col}{The column in the shape file that represents the polygons}

\item{name}{The name of the polygons}
}
\description{
Maps the shape file chosen, e.g. MPA's, and adds an annotaed map of South Africa and a line around the proposed world heritage site. The polygons within the shape file are coloured based on the column of choice, and named accordingly. The north arrow and scale bar are added to the bottown right, and a legend of the shape files polygons is added. After it has been mapped a png of the map is saved to outputs.
}
\details{
The packages tidyverse and rnaturalearth need to be installed
}
